moved MatchFixVariant into own files
