+++ src/extension/webview/components/MatchFixVariant.css
@@ -1,0 +1,31 @@
+.ms-match-fix-variant-wrapper {
+    width: 100%;
+}
+
+.ms-match-fix-variant-header-wrapper {
+
+}
+
+.ms-match-fix-variant-header-tag {
+
+}
+
+.ms-match-fix-variant-header-title {
+    flex: 1;
+}
+
+.ms-match-fix-variant-header-button-wrapper {
+    margin-right: 0.5em;
+}
+
+.ms-match-fix-variant-button-run {
+
+}
+
+.ms-match-fix-variant-content-info-statement-wrapper {
+
+}
+
+.ms-match-fix-variant-content-detail-wrapper {
+
+}
